Refactor into using a HashSet as underlying container for bukckets
The LRU needs to refer to the Keys, storing everything in a HashSet makes it easy to keep the
Key in the Entry. Furthermore the Hashtable will only store single boxed pointers which should
make this pretty efficient.

Introduces the KeyTraits trait for easier handling all the traits a key needs to
implement. Adds 'Debug' connditionally to keys depending whenever 'logging' is enabled or not.